refactor: 문서 조회 API 코드 마이독스 페이지로 통합
