+++ src/pages/MyDocsPage.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import styled from 'styled-components'
 import Header from '../components/Header'
@@ -9,6 +9,8 @@ import LanguageToggle from '../components/mydocs/upper/LanguageToggle'
 import URLInput from '../components/mydocs/upper/URLInput'
 import RoundCarousel from '../components/mydocs/upper/RoundCarousel'
 import Gallery from '../components/mydocs/lower/Gallery'
+import { Doc } from '../store/types'
+import axios from 'axios'
 
 const Container = styled.div`
   display: flex;
@@ -50,11 +52,30 @@ const MyDocsPage: React.FC = () => {
   /* 로그인 이벤트 핸들러 */
   const [isLogin, setIsLogin] = useState(true) // 기본값은 로그인이 된 상태
   const navigate = useNavigate()
-
+  const [docs, setDocs] = useState<Doc[]>([])
+  const apiUrl = 'http://gtd.kro.kr:8000/api/v1/docs/'
   const handleLogout = () => {
     setIsLogin(false) // 로그아웃
     navigate('/') // 메인페이지로 이동
   }
+
+  const getDocs = async () => {
+    try {
+      // API 호출
+      const response = await axios.get(`${apiUrl}`)
+      const docs = response.data.data.docs
+      setDocs(docs)
+      console.log(response.data.data)
+    } catch (error) {
+      // API 호출 실패
+      console.error('API Error: ', error)
+      alert('API 호출에 실패하였습니다.')
+    }
+  }
+
+  useEffect(() => {
+    getDocs()
+  }, [])
 
   /* Upper
   GiToDoc 로고 (GiToDoc)
@@ -75,10 +96,10 @@ const MyDocsPage: React.FC = () => {
           <Documentation />
           <LanguageToggle />
           <URLInput />
-          <RoundCarousel />
+          <RoundCarousel docs={docs.slice(0, 10)} />
         </Upper>
         <Lower>
-          <Gallery />
+          <Gallery docs={docs} />
         </Lower>
       </ScrollSnap>
     </Container>
