+++ src/store/types.ts
@@ -1,0 +1,6 @@
+export interface Doc {
+  id: number
+  title: string
+  updated_at: string
+  color: string
+}
